+++ libcnb-package/src/lib.rs
@@ -1,5 +1,6 @@
 // https://rust-lang.github.io/rust-clippy/stable/index.html
 #![warn(clippy::pedantic)]
+#![warn(unused_crate_dependencies)]
 // This lint is too noisy and enforces a style that reduces readability in many cases.
 #![allow(clippy::module_name_repetitions)]