+++ libcnb-test/src/build.rs
@@ -1,7 +1,7 @@
 use cargo_metadata::MetadataCommand;
-use libcnb_cargo::build::{build_buildpack_binaries, BuildBinariesError};
-use libcnb_cargo::cross_compile::{cross_compile_assistance, CrossCompileAssistance};
-use libcnb_cargo::{assemble_buildpack_directory, CargoProfile};
+use libcnb_package::build::{build_buildpack_binaries, BuildBinariesError};
+use libcnb_package::cross_compile::{cross_compile_assistance, CrossCompileAssistance};
+use libcnb_package::{assemble_buildpack_directory, CargoProfile};
 use std::path::PathBuf;
 use tempfile::{tempdir, TempDir};